requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
clipboard==0.0.4
cryptography==35.0.0
distro==1.6.0
h11==0.12.0
idna==3.3
MouseInfo==0.1.3
np==1.0.2
numpy==1.22.0
opencv-python==4.5.5.62
outcome==1.1.0
pandas==1.3.5
pdf2image==1.16.0
Pillow==8.4.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytesseract==0.3.8
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
pywin32==302
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabula-py==2.3.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
